requirements.txt
altgraph==0.17.4
appdirs==1.4.4
asyncio==3.4.3
certifi==2024.2.2
colorama==0.4.6
Deprecated==1.2.14
img2pdf==0.5.1
importlib-metadata==7.0.1
lxml==5.1.0
packaging==23.2
pefile==2023.2.7
pikepdf==8.12.0
pillow==10.2.0
pyee==8.2.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyPDF2==3.0.1
pyppeteer==1.0.2
pywin32-ctypes==0.2.2
tqdm==4.66.1
urllib3==1.26.18
websockets==10.4
wrapt==1.16.0
zipp==3.17.0
